server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
asyncpg
python-dotenv
pydantic-settings
alembic
mcp
httpx
duckduckgo-search
pyspark==3.5.3
py4j
